create paperModel

+++ src/main/java/com/tc/utils/XMLUtil.java
@@ -6,58 +6,65 @@ import org.dom4j.Element;
 import org.dom4j.io.OutputFormat;
 import org.dom4j.io.XMLWriter;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 
 /**
- *  操作XML文件的工具类
- *
+ * 操作XML文件的工具类
+ * <p>
  * Created by htc on 2017/4/21.
  */
 public class XMLUtil {
 
     /**
      * 生成试卷模板
-     *  src/main/webapp/file/paperModel/江西师大毕业考试.xml
+     * src/main/webapp/file/paperModel/江西师大毕业考试.xml
+     *
      * @param path
      * @throws Exception
      */
     public static void createPaperModel(String path) throws Exception {
-        Document doc= DocumentHelper.createDocument();
+        File file = new File(path);
+        if (file.exists()) {
+            System.out.println("文件已存在，请重新输入文件名");
+        } else {
+            Document doc = DocumentHelper.createDocument();
 
-        //生成试卷
-        Element rootElem=doc.addElement("paper");
-        //试卷头
-        Element headElem=rootElem.addElement("head");
-        headElem.addElement("id");
-        headElem.addElement("tittle");
-        headElem.addElement("score");
-        headElem.addElement("author");
-        headElem.addElement("limitTime");
-        //试卷体
-        Element bodyElem=rootElem.addElement("body");
-        bodyElem.addElement("singleChoiceQuestions");
-        bodyElem.addElement("multipleChoiceQuestions");
-        bodyElem.addElement("clozes");
-        bodyElem.addElement("judgmentQuestions");
-        bodyElem.addElement("essayQuestions");
+            //生成试卷
+            Element rootElem = doc.addElement("paper");
+            //试卷头
+            Element headElem = rootElem.addElement("head");
+            headElem.addElement("id");
+            headElem.addElement("tittle");
+            headElem.addElement("score");
+            headElem.addElement("author");
+            headElem.addElement("limitTime");
+            //试卷体
+            Element bodyElem = rootElem.addElement("body");
+            bodyElem.addElement("singleChoiceQuestions");
+            bodyElem.addElement("multipleChoiceQuestions");
+            bodyElem.addElement("clozes");
+            bodyElem.addElement("judgmentQuestions");
+            bodyElem.addElement("essayQuestions");
 
-        //指定输出路径
-        FileOutputStream fileOutputStream=new FileOutputStream(path);
-        //指定文件格式
-        OutputFormat outputFormat=OutputFormat.createPrettyPrint();
-        //设置编码
-        outputFormat.setEncoding("UTF-8");
+            //指定输出路径
+            FileOutputStream fileOutputStream = new FileOutputStream(path);
+            //指定文件格式
+            OutputFormat outputFormat = OutputFormat.createPrettyPrint();
+            //设置编码
+            outputFormat.setEncoding("UTF-8");
 
-        XMLWriter writer=new XMLWriter(fileOutputStream,outputFormat);
-        writer.write(doc);
-        writer.close();
+            XMLWriter writer = new XMLWriter(fileOutputStream, outputFormat);
+            writer.write(doc);
+            writer.close();
+        }
     }
 
-    public static void main(String[] args) throws Exception{
-        String path="src/main/webapp/file/paperModel/";
-        String fileName="江西师大毕业考试";
-//        createPaperModel(path+fileName+".xml");
-        System.out.println(path+fileName+".xml");
+    public static void main(String[] args) throws Exception {
+        String path = "src/main/webapp/file/paperModel/";
+        String fileName = "江西师大毕业考试";
+        createPaperModel(path+fileName+".xml");
+        System.out.println(path + fileName + ".xml");
     }
 }
